add auth check rest, delete error msg at error.ejs

+++ server/routes/rest.js
@@ -14,26 +14,53 @@ const user = new User(connection);
 
 /* get user router*/
 router.get("/user/:mode", (req, res) => {
-    if (req.params.mode == "all") {
-        user.getUserAll()
-            .then((result) => {
-                res.json(result);
-            })
-            .catch((err) => {
-                res.json(err);
-            });
-    } else if (req.params.mode == "withouttemp") {
-        user.getUserWithoutTemp()
-            .then((result) => {
-                res.json(result);
-            })
-            .catch((err) => {
-                res.json(err);
-            });
-    } else if (req.params.mode == "identify") {
-        let nfcid = req.query.nfcid;
-        if (nfcid) {
-            user.identify(nfcid)
+    if (req.session.auth) {
+        if (req.params.mode == "all") {
+            user.getUserAll()
+                .then((result) => {
+                    res.json(result);
+                })
+                .catch((err) => {
+                    res.json(err);
+                });
+        } else if (req.params.mode == "withouttemp") {
+            user.getUserWithoutTemp()
+                .then((result) => {
+                    res.json(result);
+                })
+                .catch((err) => {
+                    res.json(err);
+                });
+        } else if (req.params.mode == "identify") {
+            let nfcid = req.query.nfcid;
+            if (nfcid) {
+                user.identify(nfcid)
+                    .then((result) => {
+                        console.log(result);
+                        res.json(result);
+                    })
+                    .catch((err) => {
+                        console.log(err);
+                        res.json(err);
+                    });
+            } else {
+                res.json({ result: null });
+            }
+        } else {
+            res.status(404);
+            res.json({ result: null });
+        }
+    } else {
+        res.json({ result: false, msg: "권한이 없습니다." });
+    }
+});
+
+/* add user information */
+router.post("/user", (req, res) => {
+    if (req.session.auth) {
+        let target = req.body.target;
+        if (target) {
+            user.addUser(target)
                 .then((result) => {
                     console.log(result);
                     res.json(result);
@@ -43,49 +70,17 @@ router.get("/user/:mode", (req, res) => {
                     res.json(err);
                 });
         } else {
-            res.json({ result: null });
+            res.json({ result: false });
         }
     } else {
-        res.status(404);
-        res.json({ result: null });
-    }
-});
-
-/* add user information */
-router.post("/user", (req, res) => {
-    let target = req.body.target;
-    if (target) {
-        user.addUser(target)
-            .then((result) => {
-                console.log(result);
-                res.json(result);
-            })
-            .catch((err) => {
-                console.log(err);
-                res.json(err);
-            });
-    } else {
-        res.json({ result: false });
+        res.json({ result: false, msg: "권한이 없습니다." });
     }
 });
 
 /*delete user */
 router.delete("/user", (req, res) => {
-    user.deleteUser(req.body.target)
-        .then((result) => {
-            res.json(result);
-        })
-        .catch((err) => {
-            res.json(err);
-        });
-});
-
-/* add temperature information */
-router.get("/addTempData", (req, res) => {
-    let nfcid = req.query.nfcid;
-    let temperature = req.query.temperature;
-    if (nfcid && temperature) {
-        user.addTempData(nfcid, temperature)
+    if (req.session.auth) {
+        user.deleteUser(req.body.target)
             .then((result) => {
                 res.json(result);
             })
@@ -93,7 +88,28 @@ router.get("/addTempData", (req, res) => {
                 res.json(err);
             });
     } else {
-        res.json({ result: false });
+        res.json({ result: false, msg: "권한이 없습니다." });
+    }
+});
+
+/* add temperature information */
+router.get("/addTempData", (req, res) => {
+    if (req.session.auth) {
+        let nfcid = req.query.nfcid;
+        let temperature = req.query.temperature;
+        if (nfcid && temperature) {
+            user.addTempData(nfcid, temperature)
+                .then((result) => {
+                    res.json(result);
+                })
+                .catch((err) => {
+                    res.json(err);
+                });
+        } else {
+            res.json({ result: false });
+        }
+    } else {
+        res.json({ result: false, msg: "권한이 없습니다." });
     }
 });
